move infra tests from function to console
