+++ tests/Unit.Tests/Unit.Tests.csproj
@@ -31,6 +31,7 @@
     <ProjectReference Include="..\..\src\Shared\Shared.csproj" />
     <ProjectReference Include="..\..\src\SUI.Client\SUI.Client.Core\SUI.Client.Core.csproj" />
     <ProjectReference Include="..\..\src\SUI.Client\SUI.Client.StorageProcessFunction\SUI.Client.StorageProcessFunction.csproj" />
+    <ProjectReference Include="..\..\src\SUI.Client\SUI.Client.StorageProcessJob\SUI.Client.StorageProcessJob.csproj" />
     <ProjectReference Include="..\..\src\tools\dbs-response-logger\SUI.DBS.Response.Logger.Core\SUI.DBS.Response.Logger.Core.csproj" />
   </ItemGroup>
